pyproject.toml
[project]
name = "aoc2025"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "black>=25.11.0",
    "pylint==4.0.4",
    "python-dotenv==1.2.1",
    "requests>=2.32.5",
]

[tool.pylint]
disable = ["C0114", "C0115", "C0116"]
